requirements.txt
backports.csv==1.0.7
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-appconf==1.0.3
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
django-select2==7.0.3
django-suit==0.2.26
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4
mysqlclient==1.4.2.post1
oauth==1.0.1
oauth2==1.9.0.post1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==5.4.1
pycryptodome==3.7.3
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
urllib3==1.24.1
xadmin==2.0.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
